Improve and refactor hyperopt tests
